+++ src/rankitem.cpp
@@ -3,7 +3,6 @@
 #include "albert/query/rankitem.h"
 using namespace std;
 
-
-albert::RankItem::RankItem(shared_ptr<Item> i, float s):
+albert::RankItem::RankItem(shared_ptr<Item> i, double s):
     item(::move(i)), score(s)
 {}
